services/backend-for-frontend-python/requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
protobuf==4.25.3
requests==2.31.0
setuptools==69.2.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
opentelemetry-instrumentation==0.47b0
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
# honeycomb-opentelemetry==0.2.4b0
opentelemetry-api==1.26.0
opentelemetry-instrumentation-flask==0.47b0
python-dotenv==1.0.1